requirements.txt
streamlit==1.18.0
altair==4.2.2
librosa==0.9.2
soundfile==0.11.0
essentia==2.1b6.dev1110
torch==1.12.1
numpy==1.21.6
scipy==1.7.3
mido==1.3.0
pydantic==1.10.13
httpx==0.23.3
python-dotenv==0.21.1
PyYAML==6.0.1

